% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{all_ids}
\alias{all_ids}
\title{Available layer datasets}
\format{
A \code{data.frame} with 5 columns:
\describe{
\item{id}{(character) Layer id.}
\item{title}{(character) Layer title.}
\item{guid}{(integer) Layer guid.}
\item{code}{(character) Internal code, determined from tile.}
\item{departement}{(character) Layer department}
}
}
\usage{
all_ids
}
\description{
A dataset containing available layer datasets from the French
Ministry of Culture's "Atlas du Patrimoine" service.
}
\details{
Contains only data available on the 96 departments of metropolitan France.
}
\keyword{datasets}
