% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{data_check}
\alias{data_check}
\title{Check that heritage data codes are valid}
\usage{
data_check(data_code)
}
\arguments{
\item{data_code}{\code{character} vector. Data codes to check.}
}
\value{
\code{NULL}. Invisibly returns \code{NULL} if all codes are valid; otherwise throws an error.
}
\description{
Validates that provided data codes exist in the heritage layers.
}
\details{
\itemize{
\item Uses \code{get_heritage_layernames()} to get the list of valid codes.
\item Throws an error if any \code{data_code} is invalid.
}
}
\examples{
\dontrun{
data_check(c("monument", "site_archaeo"))
}

}
\keyword{internal}
