% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{data_filter}
\alias{data_filter}
\title{Filter heritage data IDs by code and department}
\usage{
data_filter(ids, department, data_code)
}
\arguments{
\item{ids}{Data frame. Input dataset of IDs. Currently ignored; function uses \code{frheritage::all_ids}.}

\item{department}{\code{character} vector. Department(s) to filter.}

\item{data_code}{\code{character} vector. Data codes to filter. Must be valid heritage layer codes.}
}
\value{
\code{data.frame}. Subset of \code{ids} filtered by \code{data_code} and \code{department}, with \code{departement} column removed.
}
\description{
Filters a dataset of heritage IDs based on provided data codes and departments.
}
\details{
\itemize{
\item Uses \code{data_check()} to ensure that all \code{data_code} values are valid.
\item The function currently ignores the \code{ids} argument and always uses \code{frheritage::all_ids}.
}
}
\examples{
\dontrun{
# Filter IDs for department "75" and code "monument"
data_filter(ids = NULL, department = "75", data_code = "monument")
}

}
\keyword{internal}
