% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo.R
\name{geo_aggregate_light}
\alias{geo_aggregate_light}
\title{Aggregate sf features by buffered centroids}
\usage{
geo_aggregate_light(x, buffer = 10)
}
\arguments{
\item{x}{An \code{sf} object (POLYGON or MULTIPOLYGON) to aggregate.}

\item{buffer}{Numeric, default \code{10}. Buffer distance (in units of the CRS) applied around centroids.}
}
\value{
An \code{sf} object with aggregated geometries as POLYGONs.
}
\description{
This internal helper function aggregates features of an \code{sf} object by:
\enumerate{
\item Computing the centroids (or central line) of the geometries.
\item Creating small buffer zones around these centroids.
\item Dissolving the buffers to generate aggregated "study zones".
}
}
\details{
\itemize{
\item The function is useful to group close features into study zones.
\item Output geometries are cast to \code{"POLYGON"} after union.
\item CRS and units are inherited from the input \code{sf} object.
}
}
\examples{
\dontrun{
library(sf)
nc <- st_read(system.file("shape/nc.shp", package = "sf"), quiet = TRUE)
zones <- geo_aggregate_light(nc, buffer = 500)
}

}
\keyword{internal}
