% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo.R
\name{geo_spatial_check}
\alias{geo_spatial_check}
\title{Validate a spatial predicate string}
\usage{
geo_spatial_check(spatial_filter)
}
\arguments{
\item{spatial_filter}{\code{character}. Spatial predicate to validate.}
}
\value{
\code{character}. The validated predicate in uppercase.
}
\description{
Checks that the provided spatial filter is one of the allowed OGC predicates.
}
\keyword{internal}
