% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ids.R
\name{ids_url_build}
\alias{ids_url_build}
\title{Build a search URL for the RSS service}
\usage{
ids_url_build(extent, insee_dep)
}
\arguments{
\item{extent}{A named list with elements \code{left}, \code{bottom}, \code{right}, \code{top}
defining the bounding box coordinates.}

\item{insee_dep}{A character or numeric INSEE department code.}
}
\value{
A character string containing the full URL to query the GeoSource service.
}
\description{
This internal helper constructs a URL to query for the French Ministry
of Culture's "Atlas du Patrimoine" GeoSource RSS service.
The URL filters data using a bounding box (\code{extent}) and an
INSEE department code (\code{insee_dep}).
}
\examples{
\dontrun{
extent <- list(left = 1, bottom = 2, right = 3, top = 4)
ids_url_build(extent, "75")
}

}
\keyword{internal}
