% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zip.R
\name{zip_download}
\alias{zip_download}
\title{Download a ZIP file from a given URL}
\usage{
zip_download(url, id)
}
\arguments{
\item{url}{\code{character}. URL to download the ZIP file from.}

\item{id}{\code{character} or \code{numeric}. Identifier of the dataset (used for messages).}
}
\value{
\code{character} or \code{NULL}.
Path to the temporary ZIP file, or \code{NULL} if download fails.
}
\description{
Downloads a ZIP file from a specified URL and saves it to a temporary file.
}
\details{
\itemize{
\item Uses \code{httr2} to perform the HTTP request.
\item If the HTTP response status is not 200, a warning is emitted and \code{NULL} is returned.
\item Errors during download are caught and a warning is printed.
}
}
\keyword{internal}
