% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ids.R
\name{ids_build_url}
\alias{ids_build_url}
\title{Build a search URL for the RSS service}
\usage{
ids_build_url(extent, insee_dep)
}
\arguments{
\item{extent}{A named list with elements \code{left}, \code{bottom}, \code{right}, \code{top}
defining the bounding box coordinates.}

\item{insee_dep}{A character or numeric INSEE department code.}
}
\value{
A character string containing the full URL to query the GeoSource service.
}
\description{
This internal helper constructs a URL to query the French Ministry
of Culture's "Atlas du Patrimoine" GeoSource service, filtered by a
bounding box (\code{extent}) and a department code (\code{insee_dep}).
}
\keyword{internal}
